% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/breslowday.R
\name{breslow_day}
\alias{breslow_day}
\title{Breslow-Day Test}
\source{
Breslow NE, Day NE. Statistical methods in cancer research. Volume I - The analysis of case-control studies. IARC Sci Publ. 1980;(32):5-338.
}
\usage{
breslow_day(x, OR = NULL, exact = FALSE)
}
\arguments{
\item{x}{a 2 by 2 by K strata array}

\item{OR}{if specified used instead of the Mantel-Haenszel}

\item{exact}{if TRUE use the hypergeometric distribution directly}
}
\description{
Breslow-Day Test
}

