% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{crossmul}
\alias{crossmul}
\title{Internal function to cross multiply a 2 x 2 table}
\usage{
crossmul(x)
}
\arguments{
\item{x}{table to cross multiply}
}
\description{
Internal function to cross multiply a 2 x 2 table
}

