% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nilton.R
\docType{data}
\name{nilton}
\alias{nilton}
\title{MRSA Dataset}
\format{A data.frame with data of 297 subjects and 6 variables:
\describe{
  \item{sex}{sex of subject, 1 = man, 0 = woman}
  \item{age}{age, in years}
  \item{methicse}{MRSA status, 1 = yes, 0 = no}
  \item{prevhosp}{previous hospitalization, 1 = yes, 0 = no}
  \item{preantbu}{previous antibiotic use, 1 = yes, 0 = no}
  \item{agecat}{Age >= 55, 1 = yes, 0 = no}
}}
\usage{
nilton
}
\description{
Data to evaluate the relationship between MRSA and prior
hopsitalization.
}
\keyword{datasets}

