% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OR.R
\name{OR}
\alias{OR}
\alias{OR.formula}
\alias{OR.table}
\title{Calculate OR}
\usage{
OR(x, ...)

\method{OR}{table}(table)

\method{OR}{formula}(formula, data)
}
\arguments{
\item{table}{a 2 x 2 table}

\item{formula}{a formula, see details}

\item{data}{data.frame containing variables formula refers to}
}
\description{
Calculate OR
}
\details{
Formula should be specified as outcome ~ exposure | strata1 + strata2...
}
\examples{
OR(prevhosp ~ methicse | agecat, nilton)
OR(prevhosp ~ methicse | agecat + preantbu, nilton)
}

