% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{strata.extract}
\alias{strata.extract}
\title{Internal function to extract a stratum from a table based on a row of expand.grid}
\usage{
strata.extract(x, tab)
}
\description{
Internal function to extract a stratum from a table based on a row of expand.grid
}

