% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{strata.name}
\alias{strata.name}
\title{Internal function that creates a strata name from row of expand.grid}
\usage{
strata.name(x)
}
\description{
Internal function that creates a strata name from row of expand.grid
}

